+++ app/shop/models/shop.go
@@ -26,7 +26,7 @@ type Shop struct {
     Amount     float64   `json:"amount" gorm:"type:double;comment:剩余金额"`
     Integral   int       `json:"integral" gorm:"type:bigint(20);comment:可用积分"`
     Tag        []ShopTag `json:"-" gorm:"many2many:shop_mark_tag;foreignKey:id;joinForeignKey:shop_id;references:id;joinReferences:tag_id;"`
-    Credit int   `gorm:"comment:授信分" json:"credit"`
+    Credit float64   `gorm:"comment:授信分" json:"credit"`
     CreateUser string    `json:"create_user" gorm:"-"`
     SalesmanUser string `json:"salesman_user" gorm:"-"`
     SalesmanPhone string `json:"salesman_phone" gorm:"-"`
